chore: restructure landing page components
