+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 import Navigation from '@/components/common/Navigation'
-import HeroSection from '@/components/common/HeroSection'
-import HowItWorksSection from '@/components/common/HowItWorksSection'
-import FeaturesSection from '@/components/common/FeaturesSection'
-import UseCasesSection from '@/components/common/UseCasesSection'
+import HeroSection from '@/components/common/landing/HeroSection'
+import HowItWorksSection from '@/components/common/landing/HowItWorksSection'
+import FeaturesSection from '@/components/common/landing/FeaturesSection'
+import UseCasesSection from '@/components/common/landing/UseCasesSection'
 import Footer from '@/components/common/Footer'
 
 export default function Home() {
