+++ src/components/common/landing/HeroSection.tsx
