+++ src/components/common/landing/HowItWorksSection.tsx
